// web/src/api-access/generated/models/shiftSeriesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Unified.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ShiftSeriesResponse {
  id?: number;
  eventSeriesId?: number;
  title?: string | null;
  description?: string | null;
  notes?: string | null;
  color?: string | null;
  recurrenceRule?: string | null;
  timeZoneId?: string | null;
  startAtUtc?: string | null;
  endAtUtc?: string | null;
  allDay?: boolean;
  eventTypeCode?: string | null;
  statusTypeCode?: string | null;
  cancelledAt?: string | null;
  cancelledByUserId?: string | null;
  cancellationReason?: string | null;
  locationId?: number | null;
  userIds?: string[];
  eventIds?: number[];
  shiftEntryIds?: number[];
}